+++ Tests/MatterControl.Tests/MatterControl/GCodeStreamTests.cs
@@ -586,9 +586,9 @@ namespace MatterControl.Tests.MatterControl
 				"; COMPLEATED_BEFORE_GCODE",
 				"; simulated after toolchange 1 gcode",
 				"G1 X9 Y8 Z7.1 F3000", // the F comes from the x movement speed
-				"G1 Z7.1 F315", // the F comes from the z movement speed
-				"G1 Z7.1 F2500", // restore the feedrate
-				"G1 Z7.1",
+				"G1 X9 Y8 Z7.1 F315", // the F comes from the z movement speed
+				"G1 X9 Y8 Z7.1 F2500", // restore the feedrate
+				"G1 X9 Y8 Z7.1",
 				// the code to switch back to t0
 				"; waiting for move on T0",
 				"",
@@ -644,9 +644,9 @@ namespace MatterControl.Tests.MatterControl
 				"; COMPLEATED_BEFORE_GCODE",
 				"; simulated after toolchange 1 gcode",
 				"G1 X9 Y8 Z7.1 F3000", // the F comes from the x movement speed
-				"G1 Z7.1 F315", // the F comes from the z movement speed
-				"G1 Z7.1 F2500", // restore the feedrate
-				"G1 Z7.1",
+				"G1 X9 Y8 Z7.1 F315", // the F comes from the z movement speed
+				"G1 X9 Y8 Z7.1 F2500", // restore the feedrate
+				"G1 X9 Y8 Z7.1",
 				// the code to switch back to t0
 				"; waiting for move on T0",
 				"",
